refactor: text encode, decode 함수 분리
